Change pbm loading logic
Previously, the pbm decoder would store values that were 1 in the
source file as a 255. This was done because it was thought that it
would make conversion to pgm files easier. This is not the case
and it is easier for the pbm logic if the values are just 0 or 1.

This fix sets what was 255 back to 1 and updates the unit tests to
work with this.